split  product and create enum
