+++ ts/components/product/product.interface.ts
@@ -1,8 +1,13 @@
+export enum productStatus {
+  available = "available",
+  outOfStock = "outOfStock",
+}
+
 export interface ProductProps {
   id: number;
   name: string;
   image: string;
   price: number;
   discount: number | null;
-  status: "available" | "outOfStock";
+  status: productStatus;
 }
