+++ ts/index.ts
@@ -1,4 +1,4 @@
-import { updateCartQty } from "./components/product/addCart.js";
+import { updateCartQty } from "./components/product/cart.function.js";
 import {
   arrivedProductsArr as arrivedProduct,
   productsDataArr as productsData,
